Refactor logging and metrics modules
