+++ backend/ai_meeting.py
@@ -8,9 +8,6 @@ import time
 import re, typing
 import textwrap
 from typing import Dict, List, Optional, Tuple
-import threading
-import csv
-import psutil
 import math
 import traceback
 import random
@@ -18,12 +15,11 @@ from pydantic import BaseModel
 from pathlib import Path
 from datetime import datetime
 
-from backend.ai_meeting.config import AgentConfig, MeetingConfig, Turn, clamp
-
-def banner(title: str):
-    print("\n" + "="*80)
-    print(f"{title}")
-    print("="*80 + "\n")
+from backend.ai_meeting.config import AgentConfig, MeetingConfig, Turn
+from backend.ai_meeting.evaluation import KPIEvaluator
+from backend.ai_meeting.logging import LiveLogWriter
+from backend.ai_meeting.metrics import MetricsLogger
+from backend.ai_meeting.utils import banner, clamp
 
 # ===== LLM Backends =====
 
@@ -81,315 +77,6 @@ class OllamaBackend(LLMBackend):
         r.raise_for_status()
         data = r.json()
         return data.get("message",{}).get("content","").strip()
-
-class LiveLogWriter:
-    def __init__(self, topic: str, outdir: Optional[str] = None, ui_minimal: bool = True):
-        ts = datetime.now().strftime("%Y%m%d-%H%M%S")
-        safe_topic = "".join(c if c.isalnum() or c in "-_（）()[]" else "_" for c in topic)[:80]
-        base_dir = Path(outdir or f"logs/{ts}_{safe_topic}")
-        base_dir.mkdir(parents=True, exist_ok=True)
-        self.dir = base_dir
-        self.md = base_dir / "meeting_live.md"
-        self.jsonl = base_dir / "meeting_live.jsonl"
-        self.html = base_dir / "meeting_live.html"
-        self.ui_minimal = ui_minimal
-        self.phase_log = base_dir / "phases.jsonl"
-        # ★ 思考ログ（デバッグ用・本文には出さない）
-        self.thoughts_log = base_dir / "thoughts.jsonl"
-
-        # ヘッダを書いておく
-        with self.md.open("w", encoding="utf-8", newline="\n") as f:
-            if self.ui_minimal:
-                f.write(f"【Topic】{topic}（開始: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}）\n\n")
-            else:
-                f.write(f"# Topic: {topic}\n\n- 開始: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}\n- ログ形式: ラウンドごとに追記\n\n")
-            f.flush()
-        # JSONLは空ファイル作成のみ
-        self.jsonl.touch()
-        self.thoughts_log.touch()
-
-    def append_turn(self, round_idx: int, turn_idx: int, speaker: str, content: str):
-        # Markdown（UI最小化＝見出しや役職を出さない）
-        line = content.strip()
-        # 箇条書き・見出しの残滓を軽く掃除
-        line = re.sub(r'^\s*[#>\-\*\u30fb・]+', '', line, flags=re.MULTILINE)
-        with self.md.open("a", encoding="utf-8", newline="\n") as f:
-            if self.ui_minimal:
-                f.write(f"{speaker}: {line}\n\n")
-            else:
-                f.write(f"{speaker}: {line}\n\n")
-            f.flush()
-        # JSONL
-        record = {
-            "ts": datetime.now().isoformat(timespec="seconds"),
-            "type": "turn",
-            "round": round_idx,
-            "turn": turn_idx,
-            "speaker": speaker,
-            "content": content,
-        }
-        with self.jsonl.open("a", encoding="utf-8", newline="\n") as f:
-            f.write(json.dumps(record, ensure_ascii=False) + "\n")
-            f.flush()
-
-    def append_phase(self, payload: Dict):
-        with self.phase_log.open("a", encoding="utf-8", newline="") as f:
-            f.write(json.dumps(payload, ensure_ascii=False) + "\n")
-            f.flush()
-
-    # ★ 全員の思考と審査結果（本文には出さない、デバッグ用）
-    def append_thoughts(self, payload: Dict):
-        with self.thoughts_log.open("a", encoding="utf-8", newline="\n") as f:
-            f.write(json.dumps(payload, ensure_ascii=False) + "\n")
-            f.flush()
-
-    # Step7: KPIコントローラの記録（本文は汚さない）
-    def append_control(self, payload: Dict):
-        with (self.dir / "control.jsonl").open("a", encoding="utf-8", newline="") as f:
-            f.write(json.dumps(payload, ensure_ascii=False) + "\n")
-            f.flush()
-
-    def append_summary(self, round_idx: int, summary: str):
-        text = summary.strip()
-        if self.ui_minimal:
-            tag = "要約"
-            with self.md.open("a", encoding="utf-8", newline="\n") as f:
-                f.write(f"（{tag}）{text}\n\n")
-                f.flush()
-        else:
-            with self.md.open("a", encoding="utf-8", newline="\n") as f:
-                f.write(f"### Round {round_idx} 要約\n\n{text}\n\n")
-                f.flush()
-        record = {
-            "ts": datetime.now().isoformat(timespec="seconds"),
-            "type": "summary",
-            "round": round_idx,
-            "summary": text,
-        }
-        with self.jsonl.open("a", encoding="utf-8", newline="\n") as f:
-            f.write(json.dumps(record, ensure_ascii=False) + "\n")
-            f.flush()
-
-    def append_final(self, final_text: str):
-        text = final_text.strip()
-        with self.md.open("a", encoding="utf-8", newline="\n") as f:
-            if self.ui_minimal:
-                f.write("【Final】\n" + text + "\n")
-            else:
-                f.write("## Final Decision / 合意案\n\n" + text + "\n")
-            f.flush()
-        record = {
-            "ts": datetime.now().isoformat(timespec="seconds"),
-            "type": "final",
-            "final": final_text,
-        }
-        with self.jsonl.open("a", encoding="utf-8", newline="\n") as f:
-            f.write(json.dumps(record, ensure_ascii=False) + "\n")
-            f.flush()
-
-    # === Step6: KPIのMarkdown追記とJSON保存 ===
-    def append_kpi(self, kpi: Dict):
-        # Markdownに追記
-        with self.md.open("a", encoding="utf-8", newline="\n") as f:
-            f.write("\n=== KPI ===\n")
-            for k,v in kpi.items():
-                f.write(f"- {k}: {v}\n")
-            f.flush()
-        # JSON保存（※ self.base_dir ではなく self.dir を使う）
-        (self.dir / "kpi.json").write_text(
-            json.dumps(kpi, ensure_ascii=False, indent=2),
-            encoding="utf-8"
-        )
-
-class KPIEvaluator:
-    def __init__(self, cfg: MeetingConfig):
-        self.cfg = cfg
-
-    def evaluate(self, history: List[Turn], pending, final_text: str) -> Dict:
-        turns = [h.content for h in history]
-        n_turns = len(turns)
-        # 1. Progress
-        if pending is not None and hasattr(pending, "items"):
-            # 初回値がなければ現在値を初期値として扱う
-            init_unres = getattr(pending, "initial", None)
-            if init_unres is None:
-                init_unres = len(pending.items)
-                setattr(pending, "initial", init_unres)
-            final_unres = len(pending.items)
-        else:
-            init_unres = 0
-            final_unres = 0
-        progress = (init_unres - final_unres) / max(1, init_unres)
-        # 2. Diversity (Jaccard近似)
-        sims = []
-        for i in range(n_turns-1):
-            a = self._token_set(turns[i]); b = self._token_set(turns[i+1])
-            sims.append(self._jacc(a,b))
-        diversity = 1 - (sum(sims)/len(sims) if sims else 0)
-        # 3. Decision density
-        decision_words = ["決定","合意","採用","実施","次回","担当","期限"]
-        hits = sum(1 for t in turns if any(w in t for w in decision_words))
-        decision_density = hits / n_turns if n_turns else 0
-        # 4. Spec coverage (Final用)
-        must = ["空間","用具","動作","得点","安全","手順","KPI"]
-        coverage = sum(1 for m in must if m in final_text) / len(must)
-        return {
-            "progress": round(progress,3),
-            "diversity": round(diversity,3),
-            "decision_density": round(decision_density,3),
-            "spec_coverage": round(coverage,3),
-        }
-
-    @staticmethod
-    def _token_set(text: str) -> set:
-        t = re.sub(r"[0-9]+"," ", text)
-        t = re.sub(r"[^\w\u3040-\u30ff\u4e00-\u9fff]+"," ", t, flags=re.UNICODE)
-        toks = [w for w in t.lower().split() if len(w)>1]
-        return set(toks)
-    @staticmethod
-    def _jacc(a:set,b:set)->float:
-        if not a or not b: return 0.0
-        return len(a&b)/len(a|b)
-
-class MetricsLogger:
-    """CPU/GPUの状態を一定間隔で記録し、終了時にPNGグラフを生成する。"""
-    def __init__(self, outdir: Path, interval: float = 1.0):
-        self.outdir = Path(outdir)
-        self.interval = max(0.5, float(interval))
-        self.csv_path = self.outdir / "metrics.csv"
-        self._stop = threading.Event()
-        self._thr = None
-        self._gpu_backend = None  # "pynvml" or "gputil" or None
-        self._init_gpu_backend()
-        # CSVヘッダ
-        with self.csv_path.open("w", encoding="utf-8", newline="") as f:
-            w = csv.writer(f)
-            w.writerow(["timestamp","cpu_percent","ram_percent","gpu_util","gpu_mem_used_mb","gpu_mem_total_mb","gpu_temp_c","gpu_power_w"])
-
-    def _init_gpu_backend(self):
-        try:
-            import pynvml
-            pynvml.nvmlInit()
-            self.nv = pynvml
-            self._gpu_backend = "pynvml"
-        except Exception:
-            try:
-                import GPUtil  # noqa
-                self._gpu_backend = "gputil"
-            except Exception:
-                self._gpu_backend = None
-
-    def _poll_gpu(self):
-        util = mem_used = mem_total = temp = power = None
-        if self._gpu_backend == "pynvml":
-            try:
-                h = self.nv.nvmlDeviceGetHandleByIndex(0)
-                util = float(self.nv.nvmlDeviceGetUtilizationRates(h).gpu)  # %
-                mem = self.nv.nvmlDeviceGetMemoryInfo(h)
-                mem_used = round(int(mem.used) / (1024*1024), 1)
-                mem_total = round(int(mem.total) / (1024*1024), 1)
-                try:
-                    temp = float(self.nv.nvmlDeviceGetTemperature(h, self.nv.NVML_TEMPERATURE_GPU))
-                except Exception:
-                    temp = None
-                try:
-                    power = round(self.nv.nvmlDeviceGetPowerUsage(h) / 1000.0, 1)  # mW→W
-                except Exception:
-                    power = None
-            except Exception:
-                pass
-        elif self._gpu_backend == "gputil":
-            try:
-                import GPUtil
-                g = GPUtil.getGPUs()[0]
-                util = float(g.load*100.0)
-                mem_used = round(g.memoryUsed, 1)
-                mem_total = round(g.memoryTotal, 1)
-                temp = float(getattr(g, "temperature", math.nan))
-                if math.isnan(temp): temp = None
-                power = None
-            except Exception:
-                pass
-        return util, mem_used, mem_total, temp, power
-
-    def _loop(self):
-        while not self._stop.is_set():
-            try:
-                ts = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-                cpu = float(psutil.cpu_percent(interval=None))
-                ram = float(psutil.virtual_memory().percent)
-                gpu_util, gpu_mu, gpu_mt, gpu_temp, gpu_pw = self._poll_gpu()
-                with self.csv_path.open("a", encoding="utf-8", newline="") as f:
-                    w = csv.writer(f)
-                    w.writerow([ts, cpu, ram, gpu_util, gpu_mu, gpu_mt, gpu_temp, gpu_pw])
-            except Exception:
-                # 例外は握りつぶして継続
-                traceback.print_exc()
-            self._stop.wait(self.interval)
-
-    def start(self):
-        self._thr = threading.Thread(target=self._loop, daemon=True)
-        self._thr.start()
-
-    def stop(self):
-        self._stop.set()
-        if self._thr:
-            self._thr.join(timeout=5)
-        # グラフ生成
-        self._make_plots()
-
-    def _make_plots(self):
-        try:
-            import matplotlib.pyplot as plt
-            import numpy as np
-            # CSV読み込み
-            rows = []
-            with self.csv_path.open("r", encoding="utf-8") as f:
-                r = csv.DictReader(f)
-                rows = list(r)
-            if not rows:
-                return
-            xs = [i for i in range(len(rows))]  # 時間軸はサンプル番号で
-            cpu = [float(r["cpu_percent"]) if r["cpu_percent"] else np.nan for r in rows]
-            ram = [float(r["ram_percent"]) if r["ram_percent"] else np.nan for r in rows]
-            gpu_util = [float(r["gpu_util"]) if r["gpu_util"] else np.nan for r in rows]
-            gpu_mu = [float(r["gpu_mem_used_mb"]) if r["gpu_mem_used_mb"] else np.nan for r in rows]
-            gpu_mt = [float(r["gpu_mem_total_mb"]) if r["gpu_mem_total_mb"] else np.nan for r in rows]
-            gpu_temp = [float(r["gpu_temp_c"]) if r["gpu_temp_c"] else np.nan for r in rows]
-            gpu_pw = [float(r["gpu_power_w"]) if r["gpu_power_w"] else np.nan for r in rows]
-
-            # CPU & RAM
-            plt.figure()
-            plt.plot(xs, cpu, label="CPU %")
-            plt.plot(xs, ram, label="RAM %")
-            plt.xlabel("samples")
-            plt.ylabel("Percent (%)")
-            plt.legend()
-            plt.title("CPU/RAM usage")
-            plt.tight_layout()
-            (self.outdir / "metrics_cpu_mem.png").unlink(missing_ok=True)
-            plt.savefig(self.outdir / "metrics_cpu_mem.png")
-            plt.close()
-
-            # GPU
-            plt.figure()
-            if any(not np.isnan(v) for v in gpu_util):
-                plt.plot(xs, gpu_util, label="GPU %")
-            if any(not np.isnan(v) for v in gpu_mu):
-                plt.plot(xs, gpu_mu, label="VRAM used (MB)")
-            if any(not np.isnan(v) for v in gpu_temp):
-                plt.plot(xs, gpu_temp, label="Temp (°C)")
-            if any(not np.isnan(v) for v in gpu_pw):
-                plt.plot(xs, gpu_pw, label="Power (W)")
-            plt.xlabel("samples")
-            plt.legend(loc='best')
-            plt.title("GPU metrics")
-            plt.tight_layout()
-            (self.outdir / "metrics_gpu.png").unlink(missing_ok=True)
-            plt.savefig(self.outdir / "metrics_gpu.png")
-            plt.close()
-        except Exception:
-            traceback.print_exc()
 
 
 # ---- Step 4: Monitor（裏方AI：フェーズ検知） ----
